* development: (48 commits)
  remove src/ for now
  Revert "v fmt and fix minor issues"
  v fmt and fix minor issues
  ...
  wisdom council
  ...
  Delete project and update test script
  project delete
  project/create...
  ...
  ...
  Implement create functions - Create functions project, issue, userstory, epic and tasks - Update post function to support auth - Create delete and edit fuctions
  fix
  fix different types conflict in redis
  List projects, issues, users, userstories, tasks and epics as objects
  taiga client
  "Use Home directory as BASEDIR
  only require threefold_data in case of gridsome websites
  parse wikis pages from proper path
  fixes
  ...

+++ redisclient/redisclient_queue.v
@@ -21,7 +21,7 @@ pub fn (mut q RedisQueue) add(val string) ? {
 
 // timeout in msec
 pub fn (mut q RedisQueue) get(timeout u64) ?string {
-	start := time.now().unix_time_milli()
+	start := u64(time.now().unix_time_milli())
 	for {
 		r := q.redis.rpop(q.key) or { '' }
 		if r != '' {
